requirements.txt
phidata
python-docx
pandas
pypdf
langchain
langchain-community
openai
python-dotenv
fastapi
uvicorn 
python-multipart
tiktoken
faiss-cpu
openpyxl
selenium